+++ lib/ui/page/home/page/chat/widget/data_attachment.dart
@@ -235,11 +235,7 @@ class _DataAttachmentState extends State<DataAttachment> {
                         mainAxisSize: MainAxisSize.min,
                         children: [
                           Text(
-                            'label_kb'.l10nfmt({
-                              'amount': e.original.size == null
-                                  ? 'dot'.l10n * 3
-                                  : e.original.size! ~/ 1024
-                            }),
+                            e.original.size.asBytes(),
                             maxLines: 1,
                             overflow: TextOverflow.ellipsis,
                             style: style.fonts.small.regular.secondary,