+++ CHANGELOG.md
@@ -18,9 +18,17 @@ All user visible changes to this project will be documented in this file. This p
         - Swipe to reply trackpad gesture. ([#1112], [#296])
         - Messages searching. ([#1116], [#692])
 
+### Changed
+
+- UI:
+    - Chat page:
+        - File sizes displayed in B, KB, MB, GB or PB. ([#1115], [#603])
+
 [#296]: /../../issues/296
+[#603]: /../../issues/603
 [#692]: /../../issues/692
 [#1112]: /../../pull/1112
+[#1115]: /../../pull/1115
 [#1116]: /../../pull/1116